Fixed `reflect.Value.Type on zero Value` panic when subscription resolver itself panicks
The internal exec Subscribe method had code to deal with subscription resolver panicking
when being called. But when such handling happen, the error is attached to the request
object and it never checked later on.

This leads to some zero checks to fail when we try to extract the type from the resolver's
channel since this variable was never set. Doing this creates a second panic which is not
handled and make the application die.

To fix the issue, we now check if there is errors on the request object before continuing
with the rest of the check, if there is errors, it's because a panic occurs and we return
the response right away.