Added possibility to subscription error properly
After a subscription has started and sent some results to the user, from the resolver
perspective, it's not right now possible to propagate error correctly to the user using the
top-level object.

Since a `chan <type>` type is required for the top-level subscription resolver, it cannot
convey any error information. By adding a `SubscriptionError` interface that top-level
resolver's object can implement, the library upon reception of the result from the channel
can check if it implements the `SubscriptionError` interface, if it's the case and the
interface's method returned a non-nil error, the subscription is terminated, the channel
is closed and the error is propagated to the user.