Added possibility to explore the actual field selection tree
Each resolver's context will now be able to access through the graph
of elements that were selected as part of the execution of this resolve via the standard
go context's value.

The returned field is an interface representing a subset of the functionality of the
internal/exec/selected package, enough for the developer to know what was queried by the
user.

I'm opening the PR to start a discussion about offering this feature to end user of the
library. We have started using it in our own fork of the library.

##### Design

This introduces new API interfaces the would become part of the standard library.

The interface is currently a simple wrapper around elements of
`internal/exec/selected/Selection` interface needed to "walk" the selected field.
Hopefully, I got the structure right.

I've currently used `Identifier()` instead of `Name()` and `Aliased()` instead of `Alias` to not renamed the field's name
since there would be conflict with struct fields in the `internal/exec/selected/Selection`
package. If we agree, the internal struct could change so the interface has the best name.

A future addition that I forsee could be:
- A "path" retrieval syntax "a la" gjson to quickly extract some nodes of the selection, for
  example with the star wars element ".search[]" that would return all type assertion nodes
  below the `search` field. Exact semantic to be discussed.
- Maybe a visitor pattern to ease the walking of the selection tree.