Redesign Directive Resolvers
The design of Directive field resolvers put burden on implementors of
directives due to the use of the `types` package. Field resolvers which
use receiver types and functions already have a more ergonomic design,
which can be repeated here for directives. This simplifies and clarifies
the implementation of directives for consumers, while putting the burden
of validating types and arguments on the library.

The implementation also incorrectly handled directive arguments
declared in the SDL, which has been corrected.