Move field resolution to fieldToExec
This helps to de-duplicate resolution logic across method vs field
resolvers, and helps prepare for changing the implementation of
field directive resolution by reducing the scope of that change.

Existing bugs in the implementation of directive resolution ('after'
only runs for fields which *cannot* return an error) are left
alone, as they'll be addressed when replacing the Before/After impl
with an alternative approach.