Loosen Directive impl Type, allowing alternate visitors
Allowing directives to not have to implement `ResolverInterceptor` in
the future, so long as they implement a valid directive visitor
function. This opens the path for directives which could execute
validation prior to any resolving of fields, without also having to
implement the resolver interceptor function.

While this opens up the type, it is still validated when parsing the
schema. When other functions like validation are supported, those
will also be accepted. This approach avoids a 'no-op' function needing
to be implemented, and the additional calls needing to be made
unnecessarily when resolving fields in the schema.

The names of types have also been updated, and types that don't need to
be exposed have been removed