Rework directives again to return some safety
Declaring an explicit interface that all directives must implement,
with function `ImplementsDirective`. This is similar to implementing
custom types, where the implementation returns `true` for the types
it implements. This allows the schema option to replace the map with
a varargs slice of directives, localising the knowledge of the name
of the directive within the implementing type.