Refactor field resolution closer to schema types
Much of the work of executing resolvers relates to the resolvable
portion of fields, rather than the selection made for the request.
With the introduction of directive visitors, these were being recreated
for each request, instead of being built once, and re-used.

By pushing the resolution of fields down into the `resolvable.Field`,
directives applied to the field definition can be setup just once when
the schema is created, and re-used across all requests.
Small exceptions to this exist for field types that embed this to
provide request specific values (resolver instance and arguments),
which get passed along with the call.

The chain of directive visitor functions still needs to be recreated
for each resolve call, because the inner resolve function needs to
accept the request specific field resolver instance, but other
per-request overheads should be removed.