Upgrade graphql-js Testdata Process
The process for pulling in test cases from graphql-js has been broken
for a long time. The project has made an array of changes which prevent
the existing approach from being functional, including migration to
Typescript, changes to assertion helper functions and more.

With this change, the approach for integration is shifted. Rather than
needing to clone the graphql-js repo, instead we pull it in as a
dependency of the Node.js project under testdata. This is patched via
patch-package in a postinstall hook, which causes the test functions to
write the cases out, and those are collected and written to JSON.

As expected, the test cases have a significant number of updates, and
the Go tests driven by them fail in their current state, so further work
is required.

Parity with test files pulled in from graphql-js prior to this change is
maintained. A number of test files were excluded due to failures, and
others have been introduced since: these are left alone for now, as it
is expected that enabling those will require larger changes to the
implementation to satisfy them.

Tests that have been added manually to the tests.json file have also been
abandoned here. The JSON file is generated entirely by the process defined,
and additional test cases will need to be recovered and re-implemented in
other ways if they are still valuable.