Refactor Validation Tests JSON
Schemas defined in the tests.json are highly dependent on the
order that graphql-js tests register them. This results in re-ordering
of the schemas and updates to many/all of the schema index values
referring to them in tests when new schemas are inserted.

Attempting to add some stability to the JSON for future changes here
by sorting the schemas, and referring to them by a SHA id of the schema
content, rather than index ordering. This results in a large change here
but hopefully smaller changes when adding and updating test cases
from graphql-js in the future.