Remove Custom Directives Support
Support for custom directives is desired to allow extension of the
capabilities and behaviour of the GraphQL server. The current
implementation would have allowed this, however the approach
has become complex and limiting. In retrospect this would be better
to be done in other ways to better support the future of the library.